+++ source/SeqFlatFileImport.Core/FileFormats/OctopusServer.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace Lib.FileFormats
+namespace SeqFlatFileImport.Core.FileFormats
 {
     public class OctopusServer : IFileFormat
     {
