+++ source/SeqFlatFileImport.Core/Reader.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
-namespace Lib
+namespace SeqFlatFileImport.Core
 {
     internal class Reader : IDisposable
     {
