+++ source/SeqFlatFileImport.Core/Result.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Lib
+namespace SeqFlatFileImport.Core
 {
     public interface IResult
     {
