+++ source/SeqFlatFileImport/SeqFlatFileImport.csproj
@@ -5,7 +5,7 @@
     <TargetFramework>net6.0</TargetFramework>
     <AssemblyName>SeqFlatFileImport</AssemblyName>
     <OutputType>Exe</OutputType>
-    <PackageId>Console</PackageId>
+    <PackageId>SeqFlatFileImport</PackageId>
     <GenerateAssemblyTitleAttribute>false</GenerateAssemblyTitleAttribute>
     <GenerateAssemblyDescriptionAttribute>false</GenerateAssemblyDescriptionAttribute>
     <GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
@@ -17,7 +17,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Lib\Lib.csproj" />
+    <ProjectReference Include="..\SeqFlatFileImport.Core\SeqFlatFileImport.Core.csproj" />
   </ItemGroup>
 
   <ItemGroup>
